django/requirements.txt
Django~=3.2.18
gunicorn~=22.0.0
psycopg2-binary~=2.8.6
djangorestframework~=3.12.4
requests~=2.31.0
graphene-django~=2.15.0
graphene-django-optimizer~=0.8.0
django-redis~=5.2.0
django-cachalot~=2.4.3
django-debug-toolbar~=3.2.2
grenml==1.0.0
openpyxl~=3.0.7
django-nested-inline~=0.4.5
django-csp~=3.7
django-q2~=1.5.1
croniter~=1.3.5
drf-spectacular~=0.26.3
importlib_resources~=5.12.0
